cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "feedub"
version = "0.1.0"
description = "Self-hosted Telegram feed reader with AI insights"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Feedub Contributors"},
]
keywords = ["telegram", "messages", "ai", "insights", "self-hosted", "llm", "aggregator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/feeduby/feedub_v2"
Repository = "https://github.com/feeduby/feedub_v2"
Issues = "https://github.com/feeduby/feedub_v2/issues"
Documentation = "https://github.com/feeduby/feedub_v2/blob/main/docs/self-hosting.md"

[project.scripts]
feedub = "feedub.main:app"

[tool.hatch.build.targets.wheel]
packages = ["feedub"]
